Cleanup some compiler warnings

+++ src/lookup_table.c
@@ -32,7 +32,7 @@
 extern int last_cap hidden;
 
 #undef cap_valid
-#define cap_valid(x) ((x) <= last_cap)
+#define cap_valid(x) ((x) <= (unsigned int)last_cap)
 
 
 struct transtab {
